+++ internal/checker/checker.go
@@ -18,7 +18,7 @@ func Check(wg *sync.WaitGroup) {
 	defer wg.Done()
 
 	lch := make(chan *types.ProxyServer, 8192)
-	probeLocal(lch)
+	probe(lch)
 	Tick(lch)
 }
 
@@ -44,14 +44,14 @@ func evictBrokenServers() {
 func Tick(lch chan<- *types.ProxyServer) {
 	//kickoff at once and repeatedly
 	evictBrokenServers()
-	queryServersForLocal(lch)
+	queryServers(lch)
 	probeTk := time.NewTicker(time.Duration(conf.Args.Probe.Interval) * time.Second)
 	evictTk := time.NewTicker(time.Duration(conf.Args.Proxy.EvictionInterval) * time.Second)
 	quit := make(chan struct{})
 	for {
 		select {
 		case <-probeTk.C:
-			queryServersForLocal(lch)
+			queryServers(lch)
 		case <-evictTk.C:
 			evictBrokenServers()
 		case <-quit:
@@ -62,7 +62,7 @@ func Tick(lch chan<- *types.ProxyServer) {
 	}
 }
 
-func queryServersForLocal(ch chan<- *types.ProxyServer) {
+func queryServers(ch chan<- *types.ProxyServer) {
 	log.Debug("collecting servers for local probe...")
 	var list []*types.ProxyServer
 	query := `SELECT 
@@ -86,15 +86,14 @@ func queryServersForLocal(ch chan<- *types.ProxyServer) {
 	}
 }
 
-func probeLocal(chjobs <-chan *types.ProxyServer) {
+func probe(chjobs <-chan *types.ProxyServer) {
 	for i := 0; i < conf.Args.Probe.Size; i++ {
 		time.Sleep(time.Millisecond * 3500)
 		go func() {
 			for ps := range chjobs {
 				var e error
 				now := util.Now()
-				if network.ValidateProxy(ps.Type, ps.Host, ps.Port,
-					conf.Args.Probe.CheckUrl, conf.Args.Probe.CheckKeyword, conf.Args.Probe.Timeout) {
+				if network.ValidateProxy(ps, conf.Args.Probe.Timeout) {
 					e = data.GormDB.Exec(`update proxy_servers set status = ?, `+
 						`suc = suc+1, score = (suc+1)/(suc+1+fail)*100, `+
 						`updated_at = ?, last_check = ? where id = ?`,