+++ internal/network/http.go
@@ -1,88 +1,90 @@
 package network
 
 import (
+	"context"
+	"crypto/tls"
 	"fmt"
 	"math/rand"
+	"net"
 	"net/http"
 	"net/url"
-	"regexp"
+	"strings"
 	"time"
 
 	"github.com/agux/roprox/internal/conf"
 	"github.com/agux/roprox/internal/types"
 	"github.com/agux/roprox/internal/ua"
+	"github.com/agux/roprox/internal/util"
 	"github.com/pkg/errors"
 	"github.com/ssgreg/repeat"
 	"golang.org/x/net/proxy"
 )
 
-const RETRY int = 3
+// TODO: need to continuously identify unrecoverable errors and stop retrying immediately.
+// Don't waste time retrying.
+// Possible situations:
+// Bad Request
+// Method Not Allowed
+// no Host in request URL
+// check whether the error message within e contains any of the key words above, ignoring case.
+var unrecoverableErrors = []string{
+	"Bad Request",
+	"Method Not Allowed",
+	"no Host in request URL",
+}
 
 // HTTPGet initiates HTTP get request and returns its response
-func HTTPGet(link string, headers map[string]string,
-	px *types.ProxyServer, cookies ...*http.Cookie) (res *http.Response, e error) {
-	host := ""
-	r := regexp.MustCompile(`//([^/]*)/`).FindStringSubmatch(link)
-	if len(r) > 0 {
-		host = r[len(r)-1]
-	}
+func HTTPGet(link string, headers map[string]string, px *types.ProxyServer, timeout, maxTimeout time.Duration,
+	cookies ...*http.Cookie) (res *http.Response, e error) {
 
-	var client *http.Client
-	req, e := http.NewRequest(http.MethodGet, link, nil)
+	//extract host from the url link
+	host := ""
+	u, err := url.Parse(link)
+	if err != nil {
+		e = errors.Wrap(err, "failed to parse URL")
+		return
+	}
+	host = u.Host
+
+	req, e := http.NewRequest(http.MethodGet, link, strings.NewReader(""))
 	if e != nil {
 		log.Panicf("unable to create http request: %+v", e)
 	}
 
-	req.Header.Set("Accept", "text/html,application/xhtml+xml,"+
-		"application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8")
-	req.Header.Set("Accept-Language", "en-US,en;q=0.9,zh-CN;q=0.8,zh;q=0.7,zh-TW;q=0.6")
-	req.Header.Set("Cache-Control", "no-cache")
-	req.Header.Set("Connection", "close")
-	if host != "" {
-		req.Header.Set("Host", host)
+	req.RequestURI = "" // Request.RequestURI can't be set in client requests
+	if req.URL.Scheme == "" {
+		req.URL.Scheme = "https"
 	}
-	req.Header.Set("Pragma", "no-cache")
-	req.Header.Set("Upgrade-Insecure-Requests", "1")
-	if headers != nil && len(headers) > 0 {
+	req.URL.Host = host
+	if req.Host == "" {
+		req.Host = host
+	}
+
+	req.Header.Set("Accept", "*/*")
+	if len(headers) > 0 {
 		for k, hv := range headers {
 			req.Header.Set(k, hv)
 		}
 	}
 	if len(req.Header.Get("User-Agent")) == 0 {
-		req.Header.Set("User-Agent", conf.Args.Network.DefaultUserAgent)
+		uaStr := conf.Args.Network.DefaultUserAgent
+		if px != nil {
+			uaStr, e = ua.GetUserAgent(px.UrlString())
+		}
+		if e != nil {
+			log.Error("failed to get user agent from the pool. falling back to default user-agent", e)
+		}
+		req.Header.Set("User-Agent", uaStr)
 	}
 
-	var proxyAddr string
-	if px == nil {
-		//no proxy used
-		client = &http.Client{Timeout: time.Second * time.Duration(conf.Args.Network.HTTPTimeout)}
-	} else {
-		proxyAddr = fmt.Sprintf("%s://%s:%s", px.Type, px.Host, px.Port)
-		switch px.Type {
-		case "socks5":
-			// create a socks5 dialer
-			dialer, e := proxy.SOCKS5("tcp", fmt.Sprintf("%s:%s", px.Host, px.Port), nil, proxy.Direct)
-			if e != nil {
-				log.Warnf("can't create socks5 proxy dialer: %+v", e)
-				return nil, errors.WithStack(e)
-			}
-			httpTransport := &http.Transport{Dial: dialer.Dial}
-			client = &http.Client{Timeout: time.Second * time.Duration(conf.Args.Network.HTTPTimeout),
-				Transport: httpTransport}
-		case "http":
-			//http proxy
-			proxyAddr := fmt.Sprintf("%s://%s:%s", px.Type, px.Host, px.Port)
-			proxyURL, e := url.Parse(proxyAddr)
-			if e != nil {
-				log.Warnf("invalid proxy: %s, %+v", proxyAddr, e)
-				return nil, errors.WithStack(e)
-			}
-			client = &http.Client{
-				Timeout:   time.Second * time.Duration(conf.Args.Network.HTTPTimeout),
-				Transport: &http.Transport{Proxy: http.ProxyURL(proxyURL)}}
-		default:
-			return nil, errors.Errorf("unsupported proxy: %+v", px)
-		}
+	var client *http.Client
+	var transport *http.Transport
+	if transport, e = GetTransport(px, true); e != nil {
+		return
+	}
+	client = &http.Client{
+		Timeout:   timeout,
+		Transport: transport,
 	}
 
 	for _, c := range cookies {
@@ -90,27 +92,35 @@ func HTTPGet(link string, headers map[string]string,
 	}
 
 	op := func(c int) error {
+		if c == 0 && px != nil {
+			log.Tracef("sending HTTP request via proxy [%s]:\n%+v", px.UrlString(), req)
+		}
 		res, e = client.Do(req)
 		if e != nil {
-			//handle "read: connection reset by peer" error by retrying
 			proxyStr := ""
-			if proxyAddr != "" {
-				proxyStr = fmt.Sprintf(" [proxy=%s]", proxyAddr)
-				UpdateProxyScore(px, false)
+			if px != nil {
+				proxyStr = fmt.Sprintf(" [proxy=%s]", px.UrlString())
 			}
-			log.Debugf("http communication error: [%+v]%s url=%s, retrying %d ...", e, proxyStr, link, c+1)
 			if res != nil {
 				res.Body.Close()
 			}
+			for _, unrecoverableError := range unrecoverableErrors {
+				if strings.Contains(strings.ToLower(e.Error()), strings.ToLower(unrecoverableError)) {
+					return repeat.HintStop(e)
+				}
+			}
+			log.Debugf("http communication error: [%+v]%s url=%s, retrying %d ...", e, proxyStr, link, c+1)
 			return repeat.HintTemporary(e)
 		}
 		return nil
 	}
 
-	e = repeat.Repeat(
+	ctx, cancel := context.WithTimeout(context.Background(), maxTimeout)
+	rp := repeat.WithContext(ctx)
+	defer cancel()
+	e = rp.Repeat(
 		repeat.FnWithCounter(op),
 		repeat.StopOnSuccess(),
-		repeat.LimitMaxTries(RETRY),
 		repeat.WithDelay(
 			repeat.FullJitterBackoff(200*time.Millisecond).WithMaxDelay(2*time.Second).Set(),
 		),
@@ -123,20 +133,20 @@ func HTTPGet(link string, headers map[string]string,
 func HTTPGetResponse(link string, headers map[string]string, useMasterProxy, rotateAgent bool) (res *http.Response, e error) {
 	host := DomainOf(link)
 	var client *http.Client
+	timeout := time.Second * time.Duration(conf.Args.Network.HTTPTimeout)
 	//determine if we must use a master proxy
 	if useMasterProxy {
-		// create a socks5 dialer
-		dialer, err := proxy.SOCKS5("tcp", conf.Args.Network.MasterProxyAddr, nil, proxy.Direct)
-		if err != nil {
-			log.Errorln("can't connect to the master proxy", err)
-			return nil, errors.WithStack(err)
+		ps := util.GetMasterProxy()
+		var transport *http.Transport
+		if transport, e = GetTransport(ps, true); e != nil {
+			return
 		}
-		// setup a http client
-		httpTransport := &http.Transport{Dial: dialer.Dial}
-		client = &http.Client{Timeout: time.Second * 60, // Maximum of 60 secs
-			Transport: httpTransport}
+		client = &http.Client{
+			Timeout:   timeout,
+			Transport: transport,
+		}
 	} else {
-		client = &http.Client{Timeout: time.Second * 60} // Maximum of 60 secs
+		client = &http.Client{Timeout: timeout}
 	}
 
 	for i := 0; true; i++ {
@@ -147,7 +157,6 @@ func HTTPGetResponse(link string, headers map[string]string, useMasterProxy, rot
 
 		req.Header.Set("Accept", "text/html,application/xhtml+xml,"+
 			"application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8")
-		req.Header.Set("Accept-Language", "en-US,en;q=0.9,zh-CN;q=0.8,zh;q=0.7,zh-TW;q=0.6")
 		req.Header.Set("Cache-Control", "no-cache")
 		req.Header.Set("Connection", "close")
 		if host != "" {
@@ -164,11 +173,10 @@ func HTTPGetResponse(link string, headers map[string]string, useMasterProxy, rot
 				continue
 			}
 		} else {
-			uagent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_1) " +
-				"AppleWebKit/537.36 (KHTML, like Gecko) Chrome/62.0.3202.94 Safari/537.36"
+			uagent = conf.Args.Network.DefaultUserAgent
 		}
 		req.Header.Set("User-Agent", uagent)
-		if headers != nil && len(headers) > 0 {
+		if len(headers) > 0 {
 			for k, hv := range headers {
 				req.Header.Set(k, hv)
 			}
@@ -189,6 +197,33 @@ func HTTPGetResponse(link string, headers map[string]string, useMasterProxy, rot
 			time.Sleep(time.Millisecond * time.Duration(500+rand.Intn(300)))
 		} else {
 			return
+		}
+	}
+	return
+}
+
+func GetTransport(ps *types.ProxyServer, insecureSkipVerify bool) (transport *http.Transport, e error) {
+	transport = &http.Transport{
+		TLSClientConfig: &tls.Config{InsecureSkipVerify: insecureSkipVerify},
+	}
+	if ps == nil {
+		return
+	}
+	if strings.HasPrefix(ps.Type, "http") {
+		var proxyURL *url.URL
+		if proxyURL, e = url.Parse(ps.UrlString()); e != nil {
+			e = errors.Wrapf(e, "Error parsing proxy URL: %s", ps.UrlString())
+			return
+		}
+		transport.Proxy = http.ProxyURL(proxyURL)
+	} else {
+		var dialer proxy.Dialer
+		if dialer, e = proxy.SOCKS5("tcp", fmt.Sprintf("%s:%s", ps.Host, ps.Port), nil, proxy.Direct); e != nil {
+			e = errors.Wrapf(e, "Error creating SOCKS5 dialer")
+			return
+		}
+		transport.DialContext = func(ctx context.Context, network, addr string) (net.Conn, error) {
+			return dialer.Dial(network, addr)
 		}
 	}
 	return