+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.3] - 2024-02-25
+
+- Switched to use outbound IP querying services to validate backend proxies, instead of configured, single website+keyword checking
+- Enhanced probe efficiency, performance, and accuracy
+
 ## [0.1.2] - 2024-02-23
 
 - Establish basic changelog and version tagging